+++ src/Prefetch.h
@@ -10,14 +10,15 @@
 #include <string>
 #include <vector>
 
-#include "Expr.h"
-#include "Schedule.h"
-#include "Target.h"
-
 namespace Halide {
+
+struct Target;
+
 namespace Internal {
 
 class Function;
+struct PrefetchDirective;
+struct Stmt;
 
 /** Inject placeholder prefetches to 's'. This placholder prefetch
   * does not have explicit region to be prefetched yet. It will be computed
