+++ src/Schedule.h
@@ -13,6 +13,7 @@
 #include "Expr.h"
 #include "FunctionPtr.h"
 #include "Parameter.h"
+#include "PrefetchDirective.h"
 
 namespace Halide {
 
@@ -85,26 +86,6 @@ enum class LoopAlignStrategy {
 
     /** By default, LoopAlignStrategy is set to NoAlign. */
     Auto
-};
-
-/** Different ways to handle accesses outside the original extents in a prefetch. */
-enum class PrefetchBoundStrategy {
-    /** Clamp the prefetched exprs by intersecting the prefetched region with
-     * the original extents. This may make the exprs of the prefetched region
-     * more complicated. */
-    Clamp,
-
-    /** Guard the prefetch with if-guards that ignores the prefetch if
-     * any of the prefetched region ever goes beyond the original extents
-     * (i.e. all or nothing). */
-    GuardWithIf,
-
-    /** Leave the prefetched exprs as are (no if-guards around the prefetch
-     * and no intersecting with the original extents). This makes the prefetch
-     * exprs simpler but this may cause prefetching of region outside the original
-     * extents. This is good if prefetch won't fault when accessing region
-     * outside the original extents. */
-    NonFaulting
 };
 
 /** A reference to a site in a Halide statement at the top of the
@@ -378,15 +359,6 @@ struct FusedPair {
         }
         return stage_2 < other.stage_2;
     }
-};
-
-struct PrefetchDirective {
-    std::string name;
-    std::string var;
-    Expr offset;
-    PrefetchBoundStrategy strategy;
-    // If it's a prefetch load from an image parameter, this points to that.
-    Parameter param;
 };
 
 struct FuncScheduleContents;
